requirements.txt
alembic==1.8.1
fastapi==0.86.0
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
SQLAlchemy==1.4.44
uvicorn==0.19.0
